resource manager is updated.

+++ trunk/ku2/ku2.c
@@ -14,27 +14,27 @@
 
 kucode_t kucode;
 unsigned int mallocs = 0;
-unsigned int sizes = 0;
+size_t sizes = 0;
 
 void printf_debug( const char *file, const char *func, int line, char *fmt, ... )
 {
 	va_list ap;
-	va_start(ap,fmt);
-	printf(">> %s:%d [%s] >> ",file,line,func);
-	vprintf(fmt,ap);
+	va_start(ap, fmt);
+	printf(">> %s:%d [%s] >> ", file, line, func);
+	vprintf(fmt, ap);
 	va_end(ap);
 }
 
-void *malloc_debug( int size )
+void *malloc_debug( size_t size )
 {
-	void *ptr=malloc(size);
-	if ( ptr!=NULL )
+	void *ptr = malloc(size);
+	if ( ptr != NULL )
 	{
 		mallocs++;
-		sizes+=size;
+		sizes += size;
 	}
 	#ifdef DEBUG_MEMORY
-	printf("MALLOC: %p sized %d (total %d)\n",ptr,size,mallocs);
+	printf("MALLOC: %p sized %d (total %d)\n", ptr, size, mallocs);
 	#endif
 	return ptr;
 }
@@ -44,6 +44,6 @@ void free_debug( void *__ptr )
 	free(__ptr);
 	mallocs--;
 	#ifdef DEBUG_MEMORY
-	printf("FREE  : %p (total left %d)\n",__ptr,mallocs);
+	printf("FREE  : %p (total left %d)\n", __ptr, mallocs);
 	#endif
 }
